refactor: Move Collection context provider to pages.
